move refresh token models to common folder
